Switching to floats. received better results.

+++ sofia/cuda/SCfinder_mem.c
@@ -14,7 +14,7 @@ double get_time(struct timeval tv1, struct timeval tv2) {
 }
 
 
-void copy3d(double *to, double *from, size_t cube_z, size_t cube_y, size_t cube_x) {
+void copy3d(float *to, float *from, size_t cube_z, size_t cube_y, size_t cube_x) {
     #pragma omp parallel for
     for (size_t z = 0; z < cube_z; z++) {
         for (size_t y = 0; y < cube_y; y++) {
@@ -30,8 +30,8 @@ void copy3d(double *to, double *from, size_t cube_z, size_t cube_y, size_t cube_
 
 
 void test_cuda(
-    double *h_in_cube,
-    double *h_out_cube,
+    float *h_in_cube,
+    float *h_out_cube,
     size_t cube_z,
     size_t cube_y,
     size_t cube_x,
@@ -50,8 +50,8 @@ void test_cuda(
 }
 
 void test_cuda_uniform(
-    double *h_in_cube,
-    double *h_out_cube,
+    float *h_in_cube,
+    float *h_out_cube,
     size_t cube_z,
     size_t cube_y,
     size_t cube_x,
@@ -68,14 +68,14 @@ void test_cuda_uniform(
 }
 
 
-void convolve_1d(double *in_cube, double *out_cube, double *weights, int cube_idx, size_t lw, int offset_multiplier, int min_clip, int max_clip) {
+void convolve_1d(float *in_cube, float *out_cube, float *weights, int cube_idx, size_t lw, int offset_multiplier, int min_clip, int max_clip) {
 
-    double sum = weights[lw] * in_cube[cube_idx];
+    float sum = weights[lw] * in_cube[cube_idx];
 
     for (size_t i = 1; i < lw + 1; i++) {
-        double lo_val = 0.0;
-        double hi_val = 0.0;
-        double weight = weights[lw + i];
+        float lo_val = 0.0;
+        float hi_val = 0.0;
+        float weight = weights[lw + i];
         int cube_offset = i * offset_multiplier;
 
         int idx_lo = cube_idx - cube_offset;
@@ -95,20 +95,20 @@ void convolve_1d(double *in_cube, double *out_cube, double *weights, int cube_id
 }
 
 
-void gaussian_filter_1d(double *in_cube, double *out_cube, size_t cube_z, size_t cube_y, size_t cube_x, size_t ks, int switch_xy) {
+void gaussian_filter_1d(float *in_cube, float *out_cube, size_t cube_z, size_t cube_y, size_t cube_x, size_t ks, int switch_xy) {
 
     int truncate = 4;
-    double sd = ((double) ks) / 2.355;
+    float sd = ((float) ks) / 2.355;
     int lw = (int) (sd * truncate + 0.5);
 
-    double* weights = (double *) malloc(sizeof(double) * (2 * lw + 1));
+    float* weights = (float *) malloc(sizeof(float) * (2 * lw + 1));
     weights[lw] = 1.0;
-    double sum = 1.0;
+    float sum = 1.0;
 
     // generate the weights
 
     for (size_t i = 1; i < lw + 1; i++) {
-        double tmp = exp(-0.5 * ((double) i * i) / (sd * sd));
+        float tmp = exp(-0.5 * ((float) i * i) / (sd * sd));
         weights[lw + i] = tmp;
         weights[lw - i] = tmp;
         sum += 2.0 * tmp;
@@ -168,7 +168,7 @@ void gaussian_filter_1d(double *in_cube, double *out_cube, size_t cube_z, size_t
 }
 
 
-void gaussian_filter(double *in_cube, double *out_cube, size_t cube_z, size_t cube_y, size_t cube_x, size_t ky, size_t kx) {
+void gaussian_filter(float *in_cube, float *out_cube, size_t cube_z, size_t cube_y, size_t cube_x, size_t ky, size_t kx) {
 
     // we recycle the cubes so as to avoid manually copying data across.
     // here in_cube ->(gaussianY)-> out_cube ->(gaussianX)-> in_cube
@@ -184,7 +184,7 @@ void gaussian_filter(double *in_cube, double *out_cube, size_t cube_z, size_t cu
 }
 
 
-void uniform_filter_1d(double *in_cube, double *out_cube, size_t cube_z, size_t cube_y, size_t cube_x, size_t kz) {
+void uniform_filter_1d(float *in_cube, float *out_cube, size_t cube_z, size_t cube_y, size_t cube_x, size_t kz) {
 
     int size1 = kz / 2;
     int size2 = kz - size1 - 1;
@@ -193,7 +193,7 @@ void uniform_filter_1d(double *in_cube, double *out_cube, size_t cube_z, size_t 
     }
 
     size_t stride = cube_x * cube_y;
-    double tmp;
+    float tmp;
 
     #pragma omp parallel for
     for (size_t x = 0; x < cube_x; x++) {
@@ -206,18 +206,18 @@ void uniform_filter_1d(double *in_cube, double *out_cube, size_t cube_z, size_t 
             for (int i = start_idx; i <= start_idx + (size2 * stride); i += stride) {
                 tmp += (i >= end_idx) ? 0.0 : in_cube[i];
             }
-            tmp /= (double) kz;
+            tmp /= (float) kz;
             out_cube[start_idx] = tmp;
 
             for (size_t z = start_idx + stride; z < end_idx; z += stride) {
                 int lo_idx = z - size1 * stride;
                 int hi_idx = z + size2 * stride;
-                double lo_val = (lo_idx < start_idx) ? 0.0 : in_cube[lo_idx];
-                double hi_val = (hi_idx >= end_idx)  ? 0.0 : in_cube[hi_idx];
-                tmp += hi_val / (double) kz;
+                float lo_val = (lo_idx < start_idx) ? 0.0 : in_cube[lo_idx];
+                float hi_val = (hi_idx >= end_idx)  ? 0.0 : in_cube[hi_idx];
+                tmp += hi_val / (float) kz;
                 out_cube[z] = tmp;
                 // remove the lower value for the next iteration
-                tmp -= lo_val / (double) kz;
+                tmp -= lo_val / (float) kz;
             }
         }
     }
@@ -261,10 +261,10 @@ void uniform_filter_1d_multi(double *in_cube, double *out_cube, size_t cube_z, s
 */
 
 // void SCfinder_mem(double *in_cube, size_t cube_z, size_t cube_y, size_t cube_x, int *kernels, size_t kern_size) {
-void SCfinder_mem(double *in_cube, size_t cube_z, size_t cube_y, size_t cube_x, int *kernel) {
+void SCfinder_mem(float *in_cube, size_t cube_z, size_t cube_y, size_t cube_x, int *kernel) {
 
     // double* out_cube = (double *) malloc(sizeof(double) * cube_x * cube_y * cube_z);
-    double* out_cube = (double *) malloc(sizeof(double) * 1);
+    float* out_cube = (float *) malloc(sizeof(float) * 1);
     // double* out_cube = init_out_cube(in_cube, cube_z, cube_y, cube_x);
 
     size_t kx = kernel[0];
